Finally

+++ srcs/tab_insert_word.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   tab_insert_word.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: root <root@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: nriker <nriker@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/03/12 09:34:04 by nriker            #+#    #+#             */
-/*   Updated: 2020/05/17 02:59:53 by root             ###   ########.fr       */
+/*   Created: 2020/08/02 05:58:45 by nriker            #+#    #+#             */
+/*   Updated: 2020/08/05 20:12:05 by nriker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,8 @@ void		curs_endofword(t_input *input)
 
 void		delete_word(t_input *input)
 {
-	while (input->line[input->x - (input->index + 1)] != ' ' && input->x > input->index
+	while (input->line[input->x - (input->index + 1)] != ' '
+			&& input->x > input->index
 			&& input->x - input->index > input->sx)
 	{
 		input->line = lineremove(input->line, input->x - input->index);
@@ -61,7 +62,7 @@ void		tab_insert_word(t_input *input, char **buf)
 	curs_endofword(input);
 	delete_word(input);
 	input->c = input->x;
-	while (x < ft_strlen(buf[0]))
+	while ((size_t)x < ft_strlen(buf[0]))
 	{
 		input->line = lineadd(&input->line, buf[0][x], input->c - input->index);
 		input->width++;
